+++ examples/deno/pagination_ts/example.ts
@@ -1,10 +1,5 @@
 import { loadSync } from "https://deno.land/std@0.173.0/dotenv/mod.ts";
-import {
-  AllowArbitraryParams,
-  config,
-  EngineParameters,
-  getJson,
-} from "../../../mod.ts";
+import { config, EngineParameters, getJson } from "../../../mod.ts";
 
 const { API_KEY: apiKey } = loadSync();
 
@@ -15,7 +10,7 @@ const params = {
   engine: "google",
   q: "Coffee",
   api_key: apiKey,
-} satisfies AllowArbitraryParams<EngineParameters<"google">>;
+} satisfies EngineParameters<"google">;
 
 // Pagination (async/await)
 let page1 = await getJson(params);